About: Quote container responsiveness WiP
